Split book generators into separate files
