+++ src/generate/gen_choicebook.cpp
@@ -1,0 +1,63 @@
+//////////////////////////////////////////////////////////////////////////
+// Purpose:   wxChoicebook generator
+// Author:    Ralph Walden
+// Copyright: Copyright (c) 2020-2022 KeyWorks Software (Ralph Walden)
+// License:   Apache License -- see ../../LICENSE
+/////////////////////////////////////////////////////////////////////////////
+
+#include <wx/choicebk.h>  // wxChoicebook: wxChoice and wxNotebook combination
+
+#include "gen_common.h"     // GeneratorLibrary -- Generator classes
+#include "gen_xrc_utils.h"  // Common XRC generating functions
+#include "node.h"           // Node class
+#include "pugixml.hpp"      // xml read/write/create/process
+#include "utils.h"          // Utility functions that work with properties
+
+#include "gen_choicebook.h"
+
+wxObject* ChoicebookGenerator::CreateMockup(Node* node, wxObject* parent)
+{
+    auto widget = new wxChoicebook(wxStaticCast(parent, wxWindow), wxID_ANY, DlgPoint(parent, node, prop_pos),
+                                   DlgSize(parent, node, prop_size), GetStyleInt(node));
+
+    widget->Bind(wxEVT_LEFT_DOWN, &BaseGenerator::OnLeftClick, this);
+    widget->Bind(wxEVT_CHOICEBOOK_PAGE_CHANGED, &ChoicebookGenerator::OnPageChanged, this);
+
+    return widget;
+}
+
+void ChoicebookGenerator::OnPageChanged(wxBookCtrlEvent& event)
+{
+    auto book = wxDynamicCast(event.GetEventObject(), wxChoicebook);
+    if (book && event.GetSelection() != wxNOT_FOUND)
+        GetMockup()->SelectNode(book->GetPage(event.GetSelection()));
+    event.Skip();
+}
+
+std::optional<ttlib::cstr> ChoicebookGenerator::GenConstruction(Node* node)
+{
+    ttlib::cstr code;
+    if (node->IsLocal())
+        code << "auto ";
+    code << node->get_node_name() << " = new wxChoicebook(";
+    code << GetParentName(node) << ", " << node->prop_as_string(prop_id);
+
+    GeneratePosSizeFlags(node, code);
+
+    return code;
+}
+
+std::optional<ttlib::cstr> ChoicebookGenerator::GenEvents(NodeEvent* event, const std::string& class_name)
+{
+    return GenEventCode(event, class_name);
+}
+
+bool ChoicebookGenerator::GetIncludes(Node* node, std::set<std::string>& set_src, std::set<std::string>& set_hdr)
+{
+    InsertGeneratorInclude(node, "#include <wx/choicebk.h>", set_src, set_hdr);
+    if (node->HasValue(prop_persist_name))
+    {
+        set_src.insert("#include <wx/persist/bookctrl.h>");
+    }
+    return true;
+}
