+++ src/generate/gen_listbook.cpp
@@ -1,0 +1,70 @@
+//////////////////////////////////////////////////////////////////////////
+// Purpose:   wxListbook generator
+// Author:    Ralph Walden
+// Copyright: Copyright (c) 2020-2022 KeyWorks Software (Ralph Walden)
+// License:   Apache License -- see ../../LICENSE
+/////////////////////////////////////////////////////////////////////////////
+
+#include <wx/listbook.h>  // wxChoicebook: wxChoice and wxNotebook combination
+
+#include "gen_book_utils.h"  // Common Book utilities
+#include "gen_common.h"      // GeneratorLibrary -- Generator classes
+#include "gen_xrc_utils.h"   // Common XRC generating functions
+#include "node.h"            // Node class
+#include "pugixml.hpp"       // xml read/write/create/process
+#include "utils.h"           // Utility functions that work with properties
+
+#include "gen_listbook.h"
+
+wxObject* ListbookGenerator::CreateMockup(Node* node, wxObject* parent)
+{
+    // Note the currently, wxListbook does not have a "style" property since the only thing that can be set is the
+    // label (tab) position
+    auto widget = new wxListbook(wxStaticCast(parent, wxWindow), wxID_ANY, DlgPoint(parent, node, prop_pos),
+                                 DlgSize(parent, node, prop_size), GetStyleInt(node));
+
+    AddBookImageList(node, widget);
+
+    widget->Bind(wxEVT_LEFT_DOWN, &BaseGenerator::OnLeftClick, this);
+    widget->Bind(wxEVT_LISTBOOK_PAGE_CHANGED, &ListbookGenerator::OnPageChanged, this);
+
+    return widget;
+}
+
+void ListbookGenerator::OnPageChanged(wxListbookEvent& event)
+{
+    auto book = wxDynamicCast(event.GetEventObject(), wxListbook);
+    if (book && event.GetSelection() != wxNOT_FOUND)
+        GetMockup()->SelectNode(book->GetPage(event.GetSelection()));
+    event.Skip();
+}
+
+std::optional<ttlib::cstr> ListbookGenerator::GenConstruction(Node* node)
+{
+    ttlib::cstr code;
+    if (node->IsLocal())
+        code << "auto ";
+    code << node->get_node_name() << " = new wxListbook(";
+    code << GetParentName(node) << ", " << node->prop_as_string(prop_id);
+
+    GeneratePosSizeFlags(node, code);
+    BookCtorAddImagelist(code, node);
+
+    return code;
+}
+
+std::optional<ttlib::cstr> ListbookGenerator::GenEvents(NodeEvent* event, const std::string& class_name)
+{
+    return GenEventCode(event, class_name);
+}
+
+bool ListbookGenerator::GetIncludes(Node* node, std::set<std::string>& set_src, std::set<std::string>& set_hdr)
+{
+    InsertGeneratorInclude(node, "#include <wx/listbook.h>", set_src, set_hdr);
+    if (node->HasValue(prop_persist_name))
+    {
+        set_src.insert("#include <wx/persist/bookctrl.h>");
+    }
+
+    return true;
+}
