+++ src/generate/gen_notebook.cpp
@@ -1,0 +1,69 @@
+//////////////////////////////////////////////////////////////////////////
+// Purpose:   wxNotebook generator
+// Author:    Ralph Walden
+// Copyright: Copyright (c) 2020-2022 KeyWorks Software (Ralph Walden)
+// License:   Apache License -- see ../../LICENSE
+/////////////////////////////////////////////////////////////////////////////
+
+#include <wx/bookctrl.h>  // wxBookCtrlBase: common base class for wxList/Tree/Notebook
+
+#include "gen_book_utils.h"  // Common Book utilities
+#include "gen_common.h"      // GeneratorLibrary -- Generator classes
+#include "gen_xrc_utils.h"   // Common XRC generating functions
+#include "node.h"            // Node class
+#include "utils.h"           // Utility functions that work with properties
+
+#include "pugixml.hpp"  // xml read/write/create/process
+
+#include "gen_notebook.h"
+
+wxObject* NotebookGenerator::CreateMockup(Node* node, wxObject* parent)
+{
+    auto widget = new wxNotebook(wxStaticCast(parent, wxWindow), wxID_ANY, DlgPoint(parent, node, prop_pos),
+                                 DlgSize(parent, node, prop_size), GetStyleInt(node));
+
+    AddBookImageList(node, widget);
+
+    widget->Bind(wxEVT_LEFT_DOWN, &BaseGenerator::OnLeftClick, this);
+    widget->Bind(wxEVT_NOTEBOOK_PAGE_CHANGED, &NotebookGenerator::OnPageChanged, this);
+
+    return widget;
+}
+
+void NotebookGenerator::OnPageChanged(wxNotebookEvent& event)
+{
+    auto book = wxDynamicCast(event.GetEventObject(), wxNotebook);
+    if (book && event.GetSelection() != wxNOT_FOUND)
+        GetMockup()->SelectNode(book->GetPage(event.GetSelection()));
+    event.Skip();
+}
+
+std::optional<ttlib::cstr> NotebookGenerator::GenConstruction(Node* node)
+{
+    ttlib::cstr code;
+    if (node->IsLocal())
+        code << "auto ";
+    code << node->get_node_name() << " = new wxNotebook(";
+    code << GetParentName(node) << ", " << node->prop_as_string(prop_id);
+
+    GeneratePosSizeFlags(node, code);
+    BookCtorAddImagelist(code, node);
+
+    return code;
+}
+
+std::optional<ttlib::cstr> NotebookGenerator::GenEvents(NodeEvent* event, const std::string& class_name)
+{
+    return GenEventCode(event, class_name);
+}
+
+bool NotebookGenerator::GetIncludes(Node* node, std::set<std::string>& set_src, std::set<std::string>& set_hdr)
+{
+    InsertGeneratorInclude(node, "#include <wx/notebook.h>", set_src, set_hdr);
+    if (node->HasValue(prop_persist_name))
+    {
+        set_src.insert("#include <wx/persist/bookctrl.h>");
+    }
+
+    return true;
+}
