+++ src/generate/gen_treebook.h
@@ -1,0 +1,26 @@
+//////////////////////////////////////////////////////////////////////////
+// Purpose:   wxTreebook generator
+// Author:    Ralph Walden
+// Copyright: Copyright (c) 2020-2022 KeyWorks Software (Ralph Walden)
+// License:   Apache License -- see ../../LICENSE
+/////////////////////////////////////////////////////////////////////////////
+
+#pragma once
+
+#include "base_generator.h"  // BaseGenerator -- Base Generator class
+
+class wxBookCtrlEvent;
+
+class TreebookGenerator : public BaseGenerator
+{
+public:
+    wxObject* CreateMockup(Node* node, wxObject* parent) override;
+
+    std::optional<ttlib::cstr> GenConstruction(Node* node) override;
+    std::optional<ttlib::cstr> GenEvents(NodeEvent* event, const std::string& class_name) override;
+
+    bool GetIncludes(Node* node, std::set<std::string>& set_src, std::set<std::string>& set_hdr) override;
+
+protected:
+    void OnPageChanged(wxBookCtrlEvent& event);
+};
